initial commit, move common functions to utils
