+++ benchmark_scrips/arctic_amplification_benchmark.py
@@ -19,7 +19,6 @@ import argparse
 import logging
 import os
 import sys
-from csv import writer
 
 import numpy as np
 import pandas as pd
@@ -28,51 +27,14 @@ from benchmark_utils import DataFinder
 
 sys.path.append("..")
 
+from utils import compute_weighted_annual_mean, save_results_csv
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, force=True)
 
 # Issue spec: Arctic must warm at least 1.5x the global mean
 ARCTIC_AMPLIFICATION_THRESHOLD = 1.5
 ARCTIC_LAT_MIN = 66.5  # degrees N
-
-
-def compute_regional_annual_mean(ds, variable, weights, lat_min=None, lat_max=None):
-    """Compute area-weighted regional-mean annual-mean time series.
-
-    Args:
-        ds: xr.Dataset containing the variable
-        variable: variable name string
-        weights: cos(lat) weights (full grid)
-        lat_min: southern latitude bound (None for global)
-        lat_max: northern latitude bound (None for global)
-
-    Returns:
-        numpy array of annual-mean regional-mean values
-    """
-    da = ds[variable]
-
-    # Apply latitude bounds if specified
-    if lat_min is not None or lat_max is not None:
-        lat = ds["lat"]
-        mask = True
-        if lat_min is not None:
-            mask = mask & (lat >= lat_min)
-        if lat_max is not None:
-            mask = mask & (lat <= lat_max)
-        da = da.where(mask, drop=True)
-        w = weights.where(mask, drop=True)
-    else:
-        w = weights
-
-    regional_mean = da.weighted(w).mean(dim=["lat", "lon"])
-
-    # Annual mean: group every 12 months (robust to non-standard calendars)
-    n_months = len(regional_mean)
-    n_years = n_months // 12
-    regional_mean = regional_mean.isel(time=slice(0, n_years * 12))
-    annual_mean = regional_mean.values.reshape(n_years, 12).mean(axis=1)
-
-    return annual_mean
 
 
 def main(
@@ -108,8 +70,8 @@ def main(
 
     # --- Compute piControl baselines (global and Arctic) ---
     logger.info("Computing piControl baselines")
-    pi_global_annual = compute_regional_annual_mean(pi_ds, "tas", weights)
-    pi_arctic_annual = compute_regional_annual_mean(
+    pi_global_annual = compute_weighted_annual_mean(pi_ds, "tas", weights)
+    pi_arctic_annual = compute_weighted_annual_mean(
         pi_ds, "tas", weights, lat_min=ARCTIC_LAT_MIN
     )
     pi_global_mean = pi_global_annual.mean()
@@ -118,8 +80,8 @@ def main(
     logger.info(f"  piControl Arctic mean: {pi_arctic_mean:.2f} K")
 
     # --- Compute abrupt-4xCO2 equilibrium response ---
-    a4x_global_annual = compute_regional_annual_mean(a4x_ds, "tas", weights)
-    a4x_arctic_annual = compute_regional_annual_mean(
+    a4x_global_annual = compute_weighted_annual_mean(a4x_ds, "tas", weights)
+    a4x_arctic_annual = compute_weighted_annual_mean(
         a4x_ds, "tas", weights, lat_min=ARCTIC_LAT_MIN
     )
 
@@ -169,36 +131,7 @@ def main(
         }
     )
 
-    if save_to_cloud:
-        from google.cloud import storage as gcs_storage
-
-        storage_client = gcs_storage.Client(project="JCM and Benchmarking")
-        bucket = storage_client.bucket("climatebench")
-        gcs_path = "results/arctic_amplification/arctic_amplification_results.csv"
-        blob = gcs_storage.Blob(bucket=bucket, name=gcs_path)
-
-        if blob.exists(storage_client):
-            import io
-
-            existing_data = blob.download_as_text()
-            output = io.StringIO(existing_data)
-            output.seek(0, io.SEEK_END)
-            writer_object = writer(output)
-            writer_object.writerow(result_df.values.flatten().tolist())
-            output.seek(0)
-            blob.upload_from_string(output.getvalue(), content_type="text/csv")
-        else:
-            result_df.to_csv(f"gs://climatebench/{gcs_path}", index=False)
-        logger.info(f"Results saved to cloud: gs://climatebench/{gcs_path}")
-    else:
-        if overwrite or not os.path.isfile(results_file):
-            os.makedirs(results_dir, exist_ok=True)
-            result_df.to_csv(results_file, index=False)
-        else:
-            with open(results_file, "a") as f:
-                writer_object = writer(f)
-                writer_object.writerow(result_df.values.flatten().tolist())
-        logger.info(f"Results saved locally: {results_file}")
+    save_results_csv(result_df, results_file, save_to_cloud, overwrite)
 
     return {
         "global_warming": global_warming,
