+++ benchmark_scrips/bjerknes_benchmark.py
@@ -28,7 +28,6 @@ import argparse
 import logging
 import os
 import sys
-from csv import writer
 
 import numpy as np
 import pandas as pd
@@ -38,11 +37,18 @@ from benchmark_utils import DataFinder
 
 sys.path.append("..")
 
+from utils import (
+    anomaly,
+    compute_meridional_transport,
+    compute_sfc_net,
+    compute_toa_net,
+    save_results_csv,
+)
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO, force=True)
 
 # --- Constants ---
-EARTH_RADIUS = 6.371e6  # metres
 COMPENSATION_LAT_MIN = 40.0  # degrees N
 COMPENSATION_LAT_MAX = 70.0  # degrees N
 DECADAL_WINDOW = 121  # months (~10 years, odd for centred rolling)
@@ -54,76 +60,9 @@ SFC_VARS = ["rsds", "rsus", "rlds", "rlus", "hfss", "hfls"]
 ALL_VARS = TOA_VARS + SFC_VARS
 
 
-# ---------------------------------------------------------------------------
-# Flux diagnostics
-# ---------------------------------------------------------------------------
-
-
-def compute_toa_net(data):
-    """TOA net downward flux (W/m2): rsdt - rsut - rlut."""
-    return data["rsdt"]["rsdt"] - data["rsut"]["rsut"] - data["rlut"]["rlut"]
-
-
-def compute_sfc_net(data):
-    """Surface net flux into ocean (W/m2).
-
-    F_sfc = (rsds - rsus) + (rlds - rlus) - hfss - hfls
-
-    Sign convention: hfss and hfls are positive upward in CMIP6, so
-    subtracting them gives the net downward flux into the surface/ocean.
-    """
-    return (
-        (data["rsds"]["rsds"] - data["rsus"]["rsus"])
-        + (data["rlds"]["rlds"] - data["rlus"]["rlus"])
-        - data["hfss"]["hfss"]
-        - data["hfls"]["hfls"]
-    )
-
-
-# ---------------------------------------------------------------------------
-# Meridional energy transport
-# ---------------------------------------------------------------------------
-
-
-def compute_meridional_transport(zonal_mean_flux, lat):
-    """Meridional energy transport by cumulative integration from the S pole.
-
-    The transport at latitude phi is the integral of the flux divergence
-    over all area south of phi:
-
-        MET(phi) = 2*pi*a^2 * integral_{-pi/2}^{phi} F(phi') cos(phi') dphi'
-
-    With discrete lat bins this becomes a cumulative sum from S to N.
-
-    Args:
-        zonal_mean_flux: xr.DataArray with dims (time, lat), in W/m2.
-        lat: latitude coordinate in degrees.
-
-    Returns:
-        xr.DataArray of meridional energy transport (W) with dims (time, lat).
-    """
-    lat_rad = np.deg2rad(lat)
-
-    # dlat for each latitude band (assume uniform spacing, take diff)
-    dlat = np.abs(np.diff(lat_rad))
-    # Pad to same length by repeating last value
-    dlat = np.append(dlat, dlat[-1])
-    dlat = xr.DataArray(dlat, dims=["lat"], coords={"lat": lat})
-
-    cos_lat = np.cos(lat_rad)
-    cos_lat = xr.DataArray(cos_lat, dims=["lat"], coords={"lat": lat})
-
-    # Integrand: F * cos(lat) * dlat * 2*pi*a^2
-    integrand = zonal_mean_flux * cos_lat * dlat * 2 * np.pi * EARTH_RADIUS**2
-
-    # Cumulative sum from south to north
-    transport = integrand.cumsum(dim="lat")
-
-    return transport
-
-
-def extract_midlat_transport(transport, lat_min=COMPENSATION_LAT_MIN,
-                             lat_max=COMPENSATION_LAT_MAX):
+def extract_midlat_transport(
+    transport, lat_min=COMPENSATION_LAT_MIN, lat_max=COMPENSATION_LAT_MAX
+):
     """Average transport over a latitude band.
 
     Args:
@@ -138,29 +77,9 @@ def extract_midlat_transport(transport, lat_min=COMPENSATION_LAT_MIN,
     return band.mean(dim="lat")
 
 
-# ---------------------------------------------------------------------------
-# Filtering and seasonal decomposition
-# ---------------------------------------------------------------------------
-
-
-def remove_seasonal_cycle(ts):
-    """Remove the climatological seasonal cycle from a monthly time series."""
-    return ts.groupby("time.month") - ts.groupby("time.month").mean("time")
-
-
 def decadal_filter(ts, window=DECADAL_WINDOW):
     """Apply a centred running-mean low-pass filter (decadal)."""
     return ts.rolling(time=window, center=True, min_periods=window // 2).mean()
-
-
-def select_djf(ts):
-    """Select December-January-February months."""
-    return ts.sel(time=ts["time.month"].isin([12, 1, 2]))
-
-
-# ---------------------------------------------------------------------------
-# Main
-# ---------------------------------------------------------------------------
 
 
 def main(
@@ -169,9 +88,7 @@ def main(
     save_to_cloud: bool = False,
     overwrite: bool = False,
 ):
-    logger.info(
-        f"Running Bjerknes compensation benchmark for {model} (piControl)"
-    )
+    logger.info(f"Running Bjerknes compensation benchmark for {model} (piControl)")
 
     # --- Load all 9 variables from piControl ---
     picontrol_data = {}
@@ -186,8 +103,7 @@ def main(
             if ensemble_members is None:
                 ensemble_members = df.ensemble_members
             logger.info(
-                f"    {var}: {len(ds.time)} months "
-                f"({len(ds.time) // 12} years)"
+                f"    {var}: {len(ds.time)} months " f"({len(ds.time) // 12} years)"
             )
         except Exception as e:
             logger.error(f"    Failed to load piControl {var}: {e}")
@@ -203,11 +119,9 @@ def main(
         )
 
     # Trim all datasets to common length
-    n_months_common = (n_years_available * 12)
+    n_months_common = n_years_available * 12
     for var in ALL_VARS:
-        picontrol_data[var] = picontrol_data[var].isel(
-            time=slice(0, n_months_common)
-        )
+        picontrol_data[var] = picontrol_data[var].isel(time=slice(0, n_months_common))
 
     # --- Compute flux fields ---
     logger.info("  Computing TOA and surface net flux fields ...")
@@ -233,17 +147,13 @@ def main(
     amet_ts = extract_midlat_transport(amet)
     omet_ts = extract_midlat_transport(omet)
 
-    logger.info(
-        f"  Mean AMET at 40-70N: {float(amet_ts.mean()) / 1e15:.3f} PW"
-    )
-    logger.info(
-        f"  Mean OMET at 40-70N: {float(omet_ts.mean()) / 1e15:.3f} PW"
-    )
+    logger.info(f"  Mean AMET at 40-70N: {float(amet_ts.mean()) / 1e15:.3f} PW")
+    logger.info(f"  Mean OMET at 40-70N: {float(omet_ts.mean()) / 1e15:.3f} PW")
 
     # --- Remove seasonal cycle and apply decadal filter ---
     logger.info("  Removing seasonal cycle and applying decadal filter ...")
-    amet_anom = remove_seasonal_cycle(amet_ts)
-    omet_anom = remove_seasonal_cycle(omet_ts)
+    amet_anom = anomaly(amet_ts)
+    omet_anom = anomaly(omet_ts)
 
     amet_decadal = decadal_filter(amet_anom)
     omet_decadal = decadal_filter(omet_anom)
@@ -258,20 +168,16 @@ def main(
         logger.error("  Insufficient data after filtering for correlation.")
         annual_corr = np.nan
     else:
-        annual_corr = float(
-            np.corrcoef(amet_filt.values, omet_filt.values)[0, 1]
-        )
+        annual_corr = float(np.corrcoef(amet_filt.values, omet_filt.values)[0, 1])
     pass_annual = annual_corr < CORRELATION_THRESHOLD
 
     logger.info(f"  Annual AMET-OMET correlation (decadal): {annual_corr:.4f}")
-    logger.info(
-        f"  Pass (r < {CORRELATION_THRESHOLD}): {pass_annual}"
-    )
+    logger.info(f"  Pass (r < {CORRELATION_THRESHOLD}): {pass_annual}")
 
     # --- DJF correlation ---
     logger.info("  Computing DJF seasonal check ...")
-    amet_djf_anom = select_djf(amet_anom)
-    omet_djf_anom = select_djf(omet_anom)
+    amet_djf_anom = amet_anom.sel(time=amet_anom["time.month"].isin([12, 1, 2]))
+    omet_djf_anom = omet_anom.sel(time=omet_anom["time.month"].isin([12, 1, 2]))
 
     amet_djf_decadal = decadal_filter(amet_djf_anom, window=DECADAL_WINDOW // 4)
     omet_djf_decadal = decadal_filter(omet_djf_anom, window=DECADAL_WINDOW // 4)
@@ -284,9 +190,7 @@ def main(
         logger.warning("  Insufficient DJF data after filtering.")
         djf_corr = np.nan
     else:
-        djf_corr = float(
-            np.corrcoef(amet_djf_filt.values, omet_djf_filt.values)[0, 1]
-        )
+        djf_corr = float(np.corrcoef(amet_djf_filt.values, omet_djf_filt.values)[0, 1])
 
     pass_djf = djf_corr < CORRELATION_THRESHOLD
     winter_stronger = (
@@ -304,9 +208,7 @@ def main(
 
     # --- Save results ---
     results_dir = "../results/bjerknes_compensation/"
-    results_file = os.path.join(
-        results_dir, "bjerknes_results.csv"
-    )
+    results_file = os.path.join(results_dir, "bjerknes_results.csv")
 
     result_df = pd.DataFrame(
         {
@@ -326,36 +228,7 @@ def main(
         }
     )
 
-    if save_to_cloud:
-        from google.cloud import storage as gcs_storage
-
-        storage_client = gcs_storage.Client(project="JCM and Benchmarking")
-        bucket = storage_client.bucket("climatebench")
-        gcs_path = "results/bjerknes_compensation/bjerknes_results.csv"
-        blob = gcs_storage.Blob(bucket=bucket, name=gcs_path)
-
-        if blob.exists(storage_client):
-            import io
-
-            existing_data = blob.download_as_text()
-            output = io.StringIO(existing_data)
-            output.seek(0, io.SEEK_END)
-            writer_object = writer(output)
-            writer_object.writerow(result_df.values.flatten().tolist())
-            output.seek(0)
-            blob.upload_from_string(output.getvalue(), content_type="text/csv")
-        else:
-            result_df.to_csv(f"gs://climatebench/{gcs_path}", index=False)
-        logger.info(f"  Results saved to cloud: gs://climatebench/{gcs_path}")
-    else:
-        if overwrite or not os.path.isfile(results_file):
-            os.makedirs(results_dir, exist_ok=True)
-            result_df.to_csv(results_file, index=False)
-        else:
-            with open(results_file, "a") as f:
-                writer_object = writer(f)
-                writer_object.writerow(result_df.values.flatten().tolist())
-        logger.info(f"  Results saved locally: {results_file}")
+    save_results_csv(result_df, results_file, save_to_cloud, overwrite)
 
     return {
         "annual_correlation": annual_corr,
