+++ utils.py
@@ -1,13 +1,17 @@
+import io
 import logging
+import os
+import time
+from csv import writer
 
 import dask.array as da
 import numpy as np
 import pandas as pd
 import requests
 import xarray as xr
+from google.cloud import storage
 
-import os
-import time
+from constants import EARTH_RADIUS
 
 logger = logging.getLogger(__name__)
 
@@ -197,3 +201,122 @@ def download_file(
             continue
 
     raise Exception(f"Failed to download after {max_retries} attempts.")
+
+
+def compute_weighted_annual_mean(ds, variable, weights, lat_min=None, lat_max=None):
+    """Compute area-weighted regional-mean annual-mean time series.
+
+    Args:
+        ds: xr.Dataset containing the variable
+        variable: variable name string
+        weights: cos(lat) weights (full grid)
+        lat_min: southern latitude bound (None for global)
+        lat_max: northern latitude bound (None for global)
+
+    Returns:
+        numpy array of annual-mean regional-mean values
+    """
+    da = ds[variable]
+
+    if lat_min is not None or lat_max is not None:
+        lat = ds["lat"]
+        mask = True
+        if lat_min is not None:
+            mask = mask & (lat >= lat_min)
+        if lat_max is not None:
+            mask = mask & (lat <= lat_max)
+        da = da.where(mask, drop=True)
+        w = weights.where(mask, drop=True)
+    else:
+        w = weights
+
+    regional_mean = da.weighted(w).mean(dim=["lat", "lon"])
+    n_months = len(regional_mean)
+    n_years = n_months // 12
+    regional_mean = regional_mean.isel(time=slice(0, n_years * 12))
+    annual_mean = regional_mean.values.reshape(n_years, 12).mean(axis=1)
+
+    return annual_mean
+
+
+def compute_toa_net(data):
+    """TOA net downward flux (W/m2): rsdt - rsut - rlut."""
+    return data["rsdt"]["rsdt"] - data["rsut"]["rsut"] - data["rlut"]["rlut"]
+
+
+def compute_sfc_net(data):
+    """Surface net flux into ocean (W/m2).
+
+    F_sfc = (rsds - rsus) + (rlds - rlus) - hfss - hfls
+
+    Sign convention: hfss and hfls are positive upward in CMIP6, so
+    subtracting them gives the net downward flux into the surface/ocean.
+    """
+    return (
+        (data["rsds"]["rsds"] - data["rsus"]["rsus"])
+        + (data["rlds"]["rlds"] - data["rlus"]["rlus"])
+        - data["hfss"]["hfss"]
+        - data["hfls"]["hfls"]
+    )
+
+
+def compute_meridional_transport(zonal_mean_flux, lat):
+    """Meridional energy transport by cumulative integration from the S pole.
+
+    MET(phi) = 2*pi*a^2 * integral_{-pi/2}^{phi} F(phi') cos(phi') dphi'
+
+    Args:
+        zonal_mean_flux: xr.DataArray with at least a 'lat' dim, in W/m2.
+        lat: latitude coordinate in degrees.
+
+    Returns:
+        xr.DataArray of meridional energy transport (W).
+    """
+    lat_rad = np.deg2rad(lat)
+
+    dlat = np.abs(np.diff(lat_rad))
+    dlat = np.append(dlat, dlat[-1])
+    dlat = xr.DataArray(dlat, dims=["lat"], coords={"lat": lat})
+
+    cos_lat = np.cos(lat_rad)
+    cos_lat = xr.DataArray(cos_lat, dims=["lat"], coords={"lat": lat})
+
+    integrand = zonal_mean_flux * cos_lat * dlat * 2 * np.pi * EARTH_RADIUS**2
+    return integrand.cumsum(dim="lat")
+
+
+def save_results_csv(result_df, results_file, save_to_cloud, overwrite):
+    """Save a results DataFrame to CSV locally or to GCS bucket "climatebench".
+
+    Args:
+        result_df: pandas DataFrame with one row of results
+        results_file: local path (e.g. "../results/ecs/ecs_results.csv")
+        save_to_cloud: if True save to GCS; False saves locally
+        overwrite: if True overwrite existing local file instead of appending
+    """
+    if save_to_cloud:
+        gcs_path = results_file[3:]  # strip "../" -> "results/benchmark/file.csv"
+        storage_client = storage.Client(project="JCM and Benchmarking")
+        bucket = storage_client.bucket("climatebench")
+        blob = storage.Blob(bucket=bucket, name=gcs_path)
+        if blob.exists(storage_client):
+            existing_data = blob.download_as_text()
+            output = io.StringIO(existing_data)
+            output.seek(0, io.SEEK_END)
+            writer_object = writer(output)
+            writer_object.writerow(result_df.values.flatten().tolist())
+            output.seek(0)
+            blob.upload_from_string(output.getvalue(), content_type="text/csv")
+        else:
+            result_df.to_csv(f"gs://climatebench/{gcs_path}", index=False)
+        logger.info(f"Results saved to cloud: gs://climatebench/{gcs_path}")
+    else:
+        results_dir = os.path.dirname(results_file)
+        if overwrite or not os.path.isfile(results_file):
+            os.makedirs(results_dir, exist_ok=True)
+            result_df.to_csv(results_file, index=False)
+        else:
+            with open(results_file, "a") as f:
+                writer_object = writer(f)
+                writer_object.writerow(result_df.values.flatten().tolist())
+        logger.info(f"Results saved locally: {results_file}")
